+++ src/components/popover/Popover.js
@@ -59,6 +59,7 @@ class Popover extends PureComponent {
       fullHeight,
       fullWidth,
       lockScroll,
+      maxWidth,
       minWidth,
       onOverlayClick,
       onEscKeyDown,
@@ -98,7 +99,7 @@ class Popover extends PureComponent {
               <div
                 data-teamleader-ui={'popover'}
                 className={cx(uiUtilities['box-shadow-200'], theme['popover'], className)}
-                style={{ left: `${left}px`, top: `${top}px`, maxWidth: fullWidth ? '100vw' : '50vw', minWidth }}
+                style={{ left: `${left}px`, top: `${top}px`, maxWidth: fullWidth ? '100vw' : maxWidth, minWidth }}
                 ref={this.popoverNode}
               >
                 <Box
@@ -149,6 +150,8 @@ Popover.propTypes = {
   fullWidth: PropTypes.bool,
   /** The scroll state of the body, if true it will not be scrollable. */
   lockScroll: PropTypes.bool,
+  /** The maximum width for the popover. */
+  maxWidth: PropTypes.string,
   /** The minimum width for the popover. */
   minWidth: PropTypes.string,
   /** The amount of extra translation on the Popover (has no effect if position is "middle" or "center"). */
@@ -180,6 +183,7 @@ Popover.defaultProps = {
   fullWidth: false,
   color: 'neutral',
   lockScroll: true,
+  maxWidth: '50vw',
   minWidth: 'auto',
   offsetCorrection: 0,
   position: 'center',